+++ src/rc.h
@@ -54,6 +54,7 @@ struct keyword {
 
 extern	int rc_parse (void);
 extern  int rc_save (void);
+extern  int rc_migrate_dir (void);
 extern  int rc_check_dir (void);